requirements.txt
absl-py==0.9.0
appdirs==1.4.3
apturl==0.5.2
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
bcrypt==3.1.7
bleach==3.1.5
blinker==1.4
Brlapi==0.7.0
cachetools==4.1.1
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2.12
django-crispy-forms==1.10.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
future==0.18.2
gast==0.3.3
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterthemes==0.20.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lesscpy==0.14.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.1
macaroonbakery==1.3.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
mysqlclient==2.0.1
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.4
notebook==6.0.3
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.7.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==11.9.0
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.4
psutil==5.5.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycryptodome==3.9.8
pycups==1.9.73
Pygments==2.6.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.16.0
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.6
s-tui==1.0.0b3
scikit-learn==0.23.1
scipy==1.4.1
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
SQLAlchemy==1.3.22
sqlparse==0.2.4
systemd-python==234
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==4.3.3
twilio==6.49.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
urwid==2.0.1
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
youtube-dl==2020.3.24
zipp==1.0.0
